Updated Term - Methods to use int for year instead of Date Updated User - Made Team relational Added methods in DAOs Updated HibernateUtil to return new openSession instead of currentSession because session is unmanaged thread-local Updated PopulatorUtil - Now populates everything Updated SQL dump - Changed term in Term from BigInt to int. (This is not the primary key)
Populator works bitches.

+++ src/main/resources/SQL/is480-scheduling.sql
@@ -55,7 +55,7 @@ CREATE TABLE `team` (
 CREATE TABLE `term` (
   `id` bigint(11) NOT NULL AUTO_INCREMENT,
   `year` date NOT NULL,
-  `term` bigint(11) NOT NULL,
+  `term` int(11) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;
